+++ src/lib/components/Spotify.svelte
@@ -2,7 +2,7 @@
   export let src: string;
 </script>
 
-<div class="mx-auto container flex flex-col lg:flex-row justify-between  px-4 mt-10 gap-4 lg:gap-20">
+<div class="mx-auto container max-w-6xl flex flex-col lg:flex-row justify-between  px-4 mt-10 gap-4 lg:gap-20">
   <iframe
   class="lg:order-1 order-2 w-full h-[400px] lg:w-3/5"
     title="Spotify Playlist"