+++ src/lib/components/Video.svelte
@@ -41,7 +41,7 @@
       >U</span
     >
   </div>
-  <div class="lg:mb-20 z-10 border border-primary lg:w-[800px]">
+  <div class="lg:mb-20 z-10 border border-primary w-full lg:w-[800px]">
     <Youtube id={videoId} animations={false} --title-color={"transparent"}>
       <button aria-label="Play" class="bg-black/50 rounded-full p-4"
         ><YTIcon class="stroke-primary w-10 h-10" /></button